make sure vars are defined

+++ jenkins/master/openshift/configuration/init.groovy.d/contra-influxdb.groovy
@@ -11,37 +11,40 @@ import jenkinsci.plugins.influxdb.models.Target
 def logger = Logger.getLogger("")
 
 env = System.getenv()
-if (env['ENABLE_INFLUXDB'].toBoolean()) {
-    logger.info('Configuring Influxdb plugin')
-    def influxdb = Jenkins.instance.getDescriptorByType(jenkinsci.plugins.influxdb.DescriptorImpl)
+if (env['ENABLE_INFLUXDB']) {
+    if (env['ENABLE_INFLUXDB'].toBoolean()) {
+        logger.info('Configuring Influxdb plugin')
+        def influxdb = Jenkins.instance.getDescriptorByType(jenkinsci.plugins.influxdb.DescriptorImpl)
 
-    try {
-        influxdb.removeTarget(env['INFLUXDB_INSTANCE'])
-        influxdb.save()
-    } finally {
-        logger.info("Cleaned up any existing localInflux targets.")
-    }
-
-    def route = env['INFLUXDB_ROUTE']
-    if (route == 'none') {
         try {
-            route = "oc get route influxdb -o jsonpath={.spec.host}".execute().text
-        } catch (err) {
-            logger.error("Failed to get influxdb route using the oc command: " + err.getMessage())
-            throw err
+            influxdb.removeTarget(env['INFLUXDB_INSTANCE'])
+            influxdb.save()
+        } finally {
+            logger.info("Cleaned up any existing localInflux targets.")
         }
+
+        def route = env['INFLUXDB_ROUTE']
+        if (route == 'none') {
+            try {
+                route = "oc get route influxdb -o jsonpath={.spec.host}".execute().text
+            } catch (err) {
+                logger.error("Failed to get influxdb route using the oc command: " + err.getMessage())
+                throw err
+            }
+        }
+
+        def target = new jenkinsci.plugins.influxdb.models.Target()
+
+        target.description = env['INFLUXDB_INSTANCE']
+        target.url = "http://${route}"
+        target.username = env['INFLUXDB_USERNAME']
+        target.password = env['INFLUXDB_PASSWORD']
+        target.database = env['INFLUXDB_DATABASE']
+
+        influxdb.addTarget(target)
+        influxdb.save()
+        logger.info("Influxdb Publisher configured")
+
     }
-
-    def target = new jenkinsci.plugins.influxdb.models.Target()
-
-    target.description = env['INFLUXDB_INSTANCE']
-    target.url = "http://${route}"
-    target.username = env['INFLUXDB_USERNAME']
-    target.password = env['INFLUXDB_PASSWORD']
-    target.database = env['INFLUXDB_DATABASE']
-
-    influxdb.addTarget(target)
-    influxdb.save()
-    logger.info("Influxdb Publisher configured")
-
 }
+
